some renaming in enum XBoxAxis/XBoxButton/XBoxStick

+++ executables/LUDUMDARE/Ludum41/src/Ludum41Game.cpp
@@ -254,10 +254,10 @@ bool LudumGame::InitializeGamepadButtonInfo()
 	if (!chaos::Game::InitializeGamepadButtonInfo())
 		return false;
 
-	gamepad_buttons.push_back(chaos::XBoxButton::BUTTON_A);
-	gamepad_buttons.push_back(chaos::XBoxButton::BUTTON_B);
-	gamepad_buttons.push_back(chaos::XBoxButton::BUTTON_X);
-	gamepad_buttons.push_back(chaos::XBoxButton::BUTTON_Y);
+	gamepad_buttons.push_back((int)chaos::XBoxButton::BUTTON_A);
+	gamepad_buttons.push_back((int)chaos::XBoxButton::BUTTON_B);
+	gamepad_buttons.push_back((int)chaos::XBoxButton::BUTTON_X);
+	gamepad_buttons.push_back((int)chaos::XBoxButton::BUTTON_Y);
 
 	return true;
 }
